Extract launching custom tabs to a new class
COAND-907
